Fixed some room placement issues
Bandaid for unit test removed, but the scenario it failed as was not
supposed to occur anyway (rooms should have two spaces apart to allow
for walls). Thus I fixed the underlying issue and even though. Thus the
unit test will still bug if rooms are only one apart, but that scenario
should never happen,  so it is theoretically fine.